Choose Active WebSecurityConfig by ConditionalOnProperty '~.auth.enabled'
